Added boolean to toggle stand ability

+++ src/commands.py
@@ -4,6 +4,7 @@ import random
 
 file = open("HelpMessage.txt", "r")
 client = discord.Client()
+standAbilityActivated = False
 
 #Outputs the current list of commands
 async def sendHelp(channel):
@@ -47,18 +48,21 @@ async def sendThunk(channel):
     print("Sent successfully")
     return;
 
+@client.event
+async def on_message(message1):
+    if(standAbilityActivated):
+
+        print("Erasing next message")
+        if message1.author == client.user:
+            return;
+            await message1.delete()
+            await channel.send(file=discord.File('./resources/EraseTime.png'))
+        standAbilityActivated = False
+    return;
 
 async def eraseTime(command, channel):
     print("Erasing command comment")
     await command.delete()
     print("Erased successfully")
-
-    @client.event
-    async def on_message(message1):
-        print("Erasing next message")
-        if message1.author == client.user:
-            return;
-        await message1.delete()
-        await channel.send(file=discord.File('./resources/EraseTime.png'))
-        return;
+    standAbilityActivated = True
     return;
